+++ samples/inbox/src/jvmMain/kotlin/com/slack/circuit/sample/inbox/main.kt
@@ -14,7 +14,6 @@ import com.slack.circuit.foundation.NavigableCircuitContent
 import com.slack.circuit.foundation.rememberCircuitNavigator
 import com.slack.circuit.retained.CircuitRetainedSettings
 import com.slack.circuit.retained.ExperimentalCircuitRetainedApi
-import com.slack.circuit.retained.RetainedValuesStoreOwner
 import com.slack.circuit.retained.RetainedValuesStoreProvider
 import com.slack.circuit.sample.inbox.di.InboxAppGraph
 import com.slack.circuit.sample.inbox.home.InboxScreen
@@ -22,20 +21,15 @@ import com.slack.circuit.sample.inbox.home.InboxScreen
 @OptIn(ExperimentalCircuitRetainedApi::class)
 fun main() {
   CircuitRetainedSettings.useFirstParty = true
-  val retainedValuesStoreOwner = RetainedValuesStoreOwner()
   val graph = InboxAppGraph.create()
   val circuitSaver = requireNotNull(graph.circuit.circuitSaver)
   application {
-    val exit = {
-      retainedValuesStoreOwner.dispose()
-      exitApplication()
-    }
     val windowState = rememberWindowState(size = DpSize(1100.dp, 800.dp))
-    Window(state = windowState, title = "Circuit Inbox", onCloseRequest = exit) {
-      RetainedValuesStoreProvider(owner = retainedValuesStoreOwner) {
+    Window(state = windowState, title = "Circuit Inbox", onCloseRequest = ::exitApplication) {
+      RetainedValuesStoreProvider {
         MaterialTheme {
           val backStack = rememberSaveableBackStack(InboxScreen, circuitSaver)
-          val navigator = rememberCircuitNavigator(backStack) { exit() }
+          val navigator = rememberCircuitNavigator(backStack) { exitApplication() }
           CircuitCompositionLocals(graph.circuit) {
             NavigableCircuitContent(navigator = navigator, backStack = backStack)
           }